[Add]server config [Fix]Customer API
