+++ api.ini
@@ -1,0 +1,30 @@
+[uwsgi]
+#application's base folder
+base = /var/www/api
+
+#python module to import
+app = wsgi
+module = %(app)
+wsgi-file = wsgi.py
+
+#the variable that holds a flask application inside the module imported at line #6
+callable = app
+
+home = %(base)
+pythonpath = %(base)
+
+#socket file's location
+socket = /var/www/api/%n.sock
+
+#permissions for the socket file
+chmod-socket = 644
+
+#location of log files
+logto = /var/log/uwsgi/%n.log
+
+master = true
+processes = 5
+
+vacuum = true
+
+die-on-term = true
