+++ app/aaa/__init__.py
@@ -2,6 +2,6 @@
 
 from flask import Blueprint
 
-api = Blueprint('api', __name__)
+aaa = Blueprint("aaa", __name__)
 
-from . import customers,errors
+from . import customers, errors
