--- app/api_1_0/decorators.py
@@ -1,16 +1,0 @@
-from functools import wraps
-
-from flask import g
-
-from app.aaa.errors import forbidden
-
-
-def permission_required(permission):
-    def decorator(f):
-        @wraps(f)
-        def decorated_function(*args, **kwargs):
-            if not g.current_user.can(permission):
-                return forbidden('Insufficient permissions')
-            return f(*args, **kwargs)
-        return decorated_function
-    return decorator
