--- app/auth/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-auth = Blueprint('auth', __name__)
-
-from . import views
