--- app/auth/forms.py
@@ -1,64 +1,0 @@
-from flask.ext.wtf import Form
-from wtforms import StringField, PasswordField, BooleanField, SubmitField
-from wtforms.validators import Required, Length, Email, Regexp, EqualTo
-from wtforms import ValidationError
-from ..models import User
-
-
-class SigninForm(Form):
-    email = StringField(label="Email", validators=[Required(), Length(1, 64),
-                                                   Email()])
-    password = PasswordField(label="Password", validators=[Required()])
-    remember_me = BooleanField('Remember me')
-    submit = SubmitField('Sign In')
-
-
-class SignupForm(Form):
-    email = StringField(label='Email', validators=[Required(), Length(1, 64),
-                                                   Email()])
-    password = PasswordField(label='Password', validators=[
-        Required(), EqualTo('password2', message='Passwords must match.')])
-    password2 = PasswordField(label='Confirm password', validators=[Required()])
-    submit = SubmitField('Sign Up')
-
-    def validate_email(self, field):
-        if User.query.filter_by(email=field.data).first():
-            raise ValidationError('Email already registered.')
-
-
-class ChangePasswordForm(Form):
-    old_password = PasswordField('Old password', validators=[Required()])
-    password = PasswordField('New password', validators=[
-        Required(), EqualTo('password2', message='Passwords must match')])
-    password2 = PasswordField('Confirm new password', validators=[Required()])
-    submit = SubmitField('Update Password')
-
-
-class PasswordResetRequestForm(Form):
-    email = StringField('Email', validators=[Required(), Length(1, 64),
-                                             Email()])
-    submit = SubmitField('Reset Password')
-
-
-class PasswordResetForm(Form):
-    email = StringField('Email', validators=[Required(), Length(1, 64),
-                                             Email()])
-    password = PasswordField('New Password', validators=[
-        Required(), EqualTo('password2', message='Passwords must match')])
-    password2 = PasswordField('Confirm password', validators=[Required()])
-    submit = SubmitField('Reset Password')
-
-    def validate_email(self, field):
-        if User.query.filter_by(email=field.data).first() is None:
-            raise ValidationError('Unknown email address.')
-
-
-class ChangeEmailForm(Form):
-    email = StringField('New Email', validators=[Required(), Length(1, 64),
-                                                 Email()])
-    password = PasswordField('Password', validators=[Required()])
-    submit = SubmitField('Update Email Address')
-
-    def validate_email(self, field):
-        if User.query.filter_by(email=field.data).first():
-            raise ValidationError('Email already registered.')
