--- app/auth/views.py
@@ -1,62 +1,0 @@
-from flask import render_template, redirect, request, url_for, flash
-from flask.ext.login import login_user, logout_user, login_required, \
-    current_user
-
-from app.aaa.email import send_email
-from . import auth
-from .forms import SigninForm, SignupForm
-from ..models import User
-
-
-@auth.before_app_request
-def before_request():
-    if current_user.is_authenticated:
-        print("current_user.is_authenticated")
-
-
-@auth.route('/signin', methods=['GET', 'POST'])
-def signin():
-    form = SigninForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user is not None and user.verify_password(form.password.data):
-            login_user(user, form.remember_me.data)
-            return redirect(request.args.get('next') or url_for('main.index'))
-        flash('Invalid username or password.')
-    return render_template('auth/signin.html', form=form)
-
-
-@auth.route('/signout')
-@login_required
-def signout():
-    logout_user()
-    flash('You have been logged out.')
-    return redirect(url_for('main.index'))
-
-
-@auth.route('/signup', methods=['GET', 'POST'])
-def signup():
-    form = SignupForm()
-    if form.validate_on_submit():
-        user = User(email=form.email.data, password=form.password.data)
-        User.add(user)
-        token = user.generate_confirmation_token()
-        send_email(user.email, 'Confirm Your Account',
-                   'auth/email/confirm', user=user, token=token)
-        flash('A confirmation email has been sent to you by email.')
-
-        login_user(user, True)
-        return redirect(url_for('main.index'))
-    return render_template('auth/signup.html', form=form)
-
-
-@auth.route('/confirm/<token>')
-@login_required
-def confirm(token):
-    if current_user.confirmed:
-        return redirect(url_for('main.index'))
-    if current_user.confirm(token):
-        flash('You have confirmed your account. Thanks!')
-    else:
-        flash('The confirmation link is invalid or has expired.')
-    return redirect(url_for('main.index'))
