+++ app/business/__init__.py
@@ -1,0 +1,4 @@
+# coding=utf-8
+
+if __name__ == '__main__':
+    pass
