+++ app/customer/__init__.py
@@ -1,0 +1,6 @@
+# coding=utf-8
+from flask import Blueprint
+
+customer = Blueprint('customer', __name__)
+
+from . import models
