--- pip-selfcheck.json
@@ -1,1 +1,0 @@
-{"last_check":"2016-05-18T15:28:19Z","pypi_version":"8.1.2"}
