+++ wsgi.py
@@ -1,0 +1,5 @@
+# coding=utf-8
+from run import app
+
+if __name__ == "__main__":
+    app.run()
